refactor: eliminate agentic-memory-helpers re-export wrapper (#593)
Move mergeExtractionConfig to agentic-memory-extraction.ts and
mergeLinkingConfig to agentic-memory-linking.ts. Update all consumers
to import directly from canonical source modules. Delete the wrapper
file and clean up vestigial comments.
